Compute diversity

Compute the diversity of a collection of recommendation lists. `Diversity' here is defined as aggregate pairwise dissimilarity, where the aggregation function and the measure of dissimilarity are specified by the user via the parameters named `agg.fun' and `method'.

Usage

diversity(
  x,
  nrec,
  method = c("jaccard", "kendall", "spearman"),
  agg.fun = mean,
  no.aggregate = FALSE,
  pred_col = "pred",
  user_col = "user",
  item_col = "item",
  ...
)

Arguments

x

data.frame that comprises a column of predictions (identified by pred_col), a column of users (identified by user_col) and a column of items (identified by item_col)

nrec

numeric scalar that specifies the length of recommendation list

method

character scalar that specifies the measure of dissimilarity to employ

agg.fun

function that aggregates the pairwise dissimilarities

no.aggregate

logical scalar that specifies to return the un-aggregated pairwise dissimilarities instead of a single aggregate value

pred_col

scalar that identifies x's column of predictions

user_col

scalar that identifies x's column of users

item_col

scalar that identifies x's column of items

...

arguments to pass to dissimilarity function

Value

numeric scalar if no.aggregate is FALSE; otherwise numeric vector of length \(n\)-choose-two, where \(n\) is the number of unique values in x[[user_col]].

See also

Examples

# Generate data.
set.seed(1L)
nuser <- 10L                        # number of unique users
nitem <- 50L                        # number of unique items
u <- paste0("u", seq_len(nuser))    # set of user IDs
i <- paste0("i", seq_len(nitem))    # set of item IDs
x <- expand.grid(pred = NA_real_, item = i, user = u)
x$pred <- unlist(replicate(nuser, { # predictions vector
    y <- runif(nitem)               # random values in [0,1]...
    y / sum(y)                      # ...that sum to one
}, simplify = FALSE))
nrec <- 5L                          # length of each recommendation list
# Compute diversity of data multiple ways for comparison.
measures <- c("jaccard", "kendall", "spearman")
names(measures) <- measures
d <- lapply(measures, function (m)
    diversity(x, nrec = nrec, method = m, no.aggregate = TRUE))
#> Error in diversity(x, nrec = nrec, method = m, no.aggregate = TRUE): object 'kendall.dist' not found
# Plot diversity values.
boxplot(d, xlab = "measure", ylab = "diversity",
    main = "Diversity by measure of dissimilarity")
#> Error in eval(expr, envir, enclos): object 'd' not found